+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Tools development version
 
+- Use miniforge3 installation on biowulf. (#124, @kelly-sovacool)
+
 ## Tools 0.4.4
 
 - Fixes for `ccbr_tools install`: (#98, @kelly-sovacool)